+++ python/pyspark/sql/tests/connect/streaming/test_parity_listener.py
@@ -16,28 +16,14 @@
 #
 
 import time
-import unittest
-from unittest import mock
 
 import pyspark.cloudpickle
-from pyspark.errors import AnalysisException, PySparkValueError
+from pyspark.errors import AnalysisException
 from pyspark.sql.functions import count, lit
 from pyspark.sql.streaming.listener import StreamingQueryListener
 from pyspark.sql.tests.streaming.test_streaming_listener import StreamingListenerTestsMixin
-from pyspark.testing.connectutils import (
-    ReusedConnectTestCase,
-    connect_requirement_message,
-    should_test_connect,
-)
+from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.utils import eventually
-
-if should_test_connect:
-    from google.protobuf import any_pb2
-
-    import pyspark.sql.connect.proto as pb2
-    from pyspark.sql.connect.client.core import PlanObservedMetrics
-    from pyspark.sql.connect.streaming.query import StreamingQueryListenerBus
-    from pyspark.sql.metrics import PlanMetrics
 
 
 # Listeners that has spark commands in callback handler functions
@@ -96,50 +82,6 @@ class TestListenerLocalV2(StreamingQueryListener):
 
     def onQueryTerminated(self, event):
         self.terminated.append(event)
-
-
-@unittest.skipIf(not should_test_connect, connect_requirement_message)
-class StreamingQueryListenerBusResponseTests(unittest.TestCase):
-    def test_ignores_execution_metadata_before_and_after_registration(self):
-        listener_added = pb2.StreamingQueryListenerEventsResult(listener_bus_listener_added=True)
-        listener_events = pb2.StreamingQueryListenerEventsResult()
-        event = listener_events.events.add(event_json="{}")
-        results = iter(
-            [
-                PlanMetrics("metric", 1, 0, []),
-                {"streaming_query_listener_events_result": listener_added},
-                PlanObservedMetrics("observation", [], []),
-                {"streaming_query_listener_events_result": listener_events},
-            ]
-        )
-        manager = mock.Mock()
-        manager._session.client.execute_command_as_iterator.return_value = results
-        listener_bus = StreamingQueryListenerBus(manager)
-
-        remaining_results = listener_bus._register_server_side_listener()
-        with (
-            mock.patch.object(
-                listener_bus, "deserialize", return_value=mock.sentinel.deserialized_event
-            ) as deserialize,
-            mock.patch.object(listener_bus, "post_to_all") as post_to_all,
-        ):
-            listener_bus._query_event_handler(remaining_results)
-
-        deserialize.assert_called_once_with(event)
-        post_to_all.assert_called_once_with(mock.sentinel.deserialized_event)
-
-    def test_rejects_unexpected_response(self):
-        responses = [
-            any_pb2.Any(),
-            {"unexpected": object()},
-            {"streaming_query_listener_events_result": object()},
-        ]
-        for response in responses:
-            with self.subTest(response=response):
-                with self.assertRaises(PySparkValueError) as error:
-                    list(StreamingQueryListenerBus._iter_listener_events(iter([response])))
-
-                self.assertEqual(error.exception.getCondition(), "UNKNOWN_RESPONSE")
 
 
 class StreamingListenerParityTests(StreamingListenerTestsMixin, ReusedConnectTestCase):
